deepl-wiki-backend/requirements.txt
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
requests>=2.28.0
uvicorn>=0.20.0
fastapi>=0.95.0
langgraph>=0.0.55
langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0
gitpython>=3.1.0
fastapi>=0.100.0